market_api_using_data_gov/agri_market_app/backend/requirements.txt
Flask>=2.0
requests>=2.25
python-dotenv>=0.21.0